move LockedRangeList into its own file for easier include
