+++ src/coreclr/src/vm/stubmgr.h
@@ -47,6 +47,7 @@
 #define __stubmgr_h__
 
 #include "simplerwlock.hpp"
+#include "lockedrangelist.h"
 
 // When 'TraceStub' returns, it gives the address of where the 'target' is for a stub'
 // TraceType indicates what this 'target' is
@@ -337,53 +338,6 @@ private:
 
     static CrstStatic s_StubManagerListCrst;
 #endif // !CROSSGEN_COMPILE
-};
-
-// -------------------------------------------------------
-// This just wraps the RangeList methods in a read or
-// write lock depending on the operation.
-// -------------------------------------------------------
-
-class LockedRangeList : public RangeList
-{
-  public:
-    VPTR_VTABLE_CLASS(LockedRangeList, RangeList)
-
-    LockedRangeList() : RangeList(), m_RangeListRWLock(COOPERATIVE_OR_PREEMPTIVE, LOCK_TYPE_DEFAULT)
-    {
-        LIMITED_METHOD_CONTRACT;
-    }
-
-    ~LockedRangeList()
-    {
-        LIMITED_METHOD_CONTRACT;
-    }
-
-  protected:
-
-    virtual BOOL AddRangeWorker(const BYTE *start, const BYTE *end, void *id)
-    {
-        WRAPPER_NO_CONTRACT;
-        SimpleWriteLockHolder lh(&m_RangeListRWLock);
-        return RangeList::AddRangeWorker(start,end,id);
-    }
-
-    virtual void RemoveRangesWorker(void *id, const BYTE *start = NULL, const BYTE *end = NULL)
-    {
-        WRAPPER_NO_CONTRACT;
-        SimpleWriteLockHolder lh(&m_RangeListRWLock);
-        RangeList::RemoveRangesWorker(id,start,end);
-    }
-
-    virtual BOOL IsInRangeWorker(TADDR address, TADDR *pID = NULL)
-    {
-        WRAPPER_NO_CONTRACT;
-        SUPPORTS_DAC;
-        SimpleReadLockHolder lh(&m_RangeListRWLock);
-        return RangeList::IsInRangeWorker(address, pID);
-    }
-
-    SimpleRWLock m_RangeListRWLock;
 };
 
 #ifndef CROSSGEN_COMPILE
